+++ templates/base.html.twig
@@ -10,6 +10,7 @@
             <a href="{{ url('categories') }}">Categories</a>
             <a href="{{ url('products') }}">Products</a>
             <a href="{{ url('users') }}">Users</a>
+            <a href="{{ url('favorites') }}">Favorites</a>
         </nav>
     </header>
     <body>
